[REFACTOR] 헤더뷰, 새로고침 분리
